+++ resources/lang/en/resource-type-item/parameters.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'title-include-categories' => 'Include assigned category?',
+    'description-include-categories' => 'Optionally include the category assigned to each item',
+
+    'title-include-subcategories' => 'Include assigned subcategory?',
+    'description-include-subcategories' => 'Optionally include the subcategory assigned to each item'
+];