+++ routes/api/public-routes.php
@@ -99,6 +99,11 @@ Route::group(
             'ResourceTypeItemController@index'
         );
 
+        Route::options(
+            'resource-types/{resource_type_id}/items',
+            'ResourceTypeItemController@optionsIndex'
+        );
+
         Route::get(
             'resource-types/{resource_type_id}/resources',
             'ResourceController@index'