+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 The complete changelog for the Costs to Expect REST API, follows the format defined at https://keepachangelog.com/en/1.0.0/
 
 ## [v1.13.0] - 2019-xx-xx
+### Added 
+- Added new route /resource-types/[resource-type]/items
+
 ### Changed
 - Updated the API landing page to point to the new website.
 - Updated the README to include the new website URL.